gltfpack: Move WASI stubs to wasistubs.cpp
